+++ src/sna/ActuatorAvAnimator.ts
@@ -1,8 +1,10 @@
 import { ActProperties } from "./SNA";
 import { ActuatorAbstract } from "./SNA";
 import { SNAManager } from "./SNA";
-import { AnimationRange, Animatable, ArcRotateCamera, Mesh, Skeleton, Scene, Vector3 } from "babylonjs";
+import { AnimationRange, Animatable, ArcRotateCamera, Mesh, Skeleton, Scene, Vector3, AnimationGroup } from "babylonjs";
 import { SelectType } from "../gui/VishvaGUI";
+import { AnimUtils } from "../util/AnimUtils";
+import { Vishva } from "../Vishva";
 
 export class AvAnimatorProp extends ActProperties {
     changeTrans: boolean = true;
@@ -11,7 +13,7 @@ export class AvAnimatorProp extends ActProperties {
     makeChild: boolean = true;
     focusOnAV: boolean = true;
     focusPosition: Vector3 = new Vector3(0, 0, 0);
-    animationRange: SelectType = new SelectType();
+    animation: SelectType = new SelectType();
     rate: number = 1;
 }
 
@@ -27,6 +29,9 @@ export class AvAnimatorProp extends ActProperties {
  */
 export class ActuatorAvAnimator extends ActuatorAbstract {
 
+    _isAG:boolean = false;
+    _skel:Skeleton = null;
+
     public constructor(mesh: Mesh, parms: AvAnimatorProp) {
         super(mesh, parms != null ? parms : new AvAnimatorProp());
 
@@ -37,23 +42,47 @@ export class ActuatorAvAnimator extends ActuatorAbstract {
         this._inControl = false;
 
         var prop: AvAnimatorProp = <AvAnimatorProp>this.properties;
-        var scene: Scene = this.mesh.getScene();
-        let avMesh = scene.getMeshesByTags("Vishva.avatar")[0];
-        var skel: Skeleton = avMesh.skeleton;
-        if (skel != null) {
-            var ranges: AnimationRange[] = skel.getAnimationRanges();
-            var animNames: string[] = new Array(ranges.length);
-            var i: number = 0;
-            for (let range of ranges) {
-                animNames[i] = range.name;
-                i++;
-            }
-            prop.animationRange.values = animNames;
-        } else {
-            prop.animationRange.values = [""];
-        }
+        let avMesh = Vishva.vishva.avatar;
+        let sm = AnimUtils.getMeshSkel(avMesh, true);
+        this._skel = (sm === null) ? null : sm.skel;
+        prop.animation.values = [""];
+        if (this._skel != null) {
+                    //check for animation ranges and animation groups
+                    let ranges: AnimationRange[] = this._skel.getAnimationRanges();
+                    if (ranges.length != 0) {
+                        let animNames: string[] = new Array();
+                        let i: number = 0;
+                        for (let range of ranges) {
+                            if (range != null) {
+                                animNames[i] = range.name;
+                                i++;
+                            }
+                        }
+                        prop.animation.values = animNames;
+                        this._skel.enableBlending(0.05);
+                    }else{
+                       let ags: AnimationGroup[] = AnimUtils.getMeshAg(avMesh,Vishva.vishva.scene.animationGroups,true);
+                       if(ags.length != 0){
+                            this._isAG = true;
+                           let animNames: string[] = new Array();
+                           let i: number = 0;
+                           for (let ag of ags) {
+                               if (ag != null) {
+                                   animNames[i] = ag.name;
+                                   i++;
+                               }
+                           }
+                           prop.animation.values = animNames;
+                           this._skel.enableBlending(0.05);
+                       }
+                    }
+                } else{
+                    console.log("no skeleton");
+                }
     }
     private anim: Animatable;
+    private ag:AnimationGroup;
+
     private avMesh: Mesh;
     //save AV position, rotation
     private _sp: Vector3;
@@ -65,14 +94,13 @@ export class ActuatorAvAnimator extends ActuatorAbstract {
     private _inControl: boolean;
 
     public actuate() {
-        console.log("actuating!!");
         if (this._inControl) return;
 
         this._inControl = true;
         let prop: AvAnimatorProp = <AvAnimatorProp>this.properties;
         this.avMesh = SNAManager.getSNAManager().getAV();
-        let skel: Skeleton = this.avMesh.skeleton;
-        if (skel != null) {
+        //let skel: Skeleton = this.avMesh.skeleton;
+        if (this._skel != null) {
             SNAManager.getSNAManager().disableAV();
 
             this._sp.copyFrom(this.avMesh.position);
@@ -101,7 +129,15 @@ export class ActuatorAvAnimator extends ActuatorAbstract {
                 //camera.target.copyFrom(this.avMesh.position);
             }
 
-            this.anim = skel.beginAnimation(prop.animationRange.value, prop.loop, prop.rate);
+            //this.anim = skel.beginAnimation(prop.animation.value, prop.loop, prop.rate);
+            if (this._skel != null) {
+            if(this._isAG){
+                this.ag = Vishva.vishva.scene.getAnimationGroupByName(prop.animation.value).start(prop.loop, prop.rate);
+                this.ag.onAnimationGroupEndObservable.addOnce(() => { return this.onActuateEnd() });
+            }else{
+                this.anim = this.avMesh.skeleton.beginAnimation(prop.animation.value, prop.loop, prop.rate,() => { return this.onActuateEnd() });
+            }
+        }
         }
     }
 
@@ -109,7 +145,12 @@ export class ActuatorAvAnimator extends ActuatorAbstract {
         if (!this._inControl) return;
         let prop: AvAnimatorProp = <AvAnimatorProp>this.properties;
         //anim would be null if user deletes the actuator without it ever being actuated
-        if (this.anim != null) this.anim.stop();
+        //if (this.anim != null) this.anim.stop();
+        if(this._isAG){
+            this.ag.stop();
+        }else{
+            if (this.anim != null) this.anim.stop();
+        }
         this.avMesh.parent = null;
         this.avMesh.position.copyFrom(this._sp);
         this.avMesh.rotation.copyFrom(this._sr);