+++ src/sna/SensorContact.ts
@@ -1,3 +1,5 @@
+import { AnimUtils } from "../util/AnimUtils";
+import { Vishva } from "../Vishva";
 import { SNAproperties } from "./SNA";
 import { SensorAbstract } from "./SNA";
 import { SNAManager } from "./SNA";
@@ -50,7 +52,13 @@ export class SensorContact extends SensorAbstract {
             this.mesh.actionManager = new ActionManager(scene);
         }
 
-        let otherMesh = scene.getMeshesByTags("Vishva.avatar")[0];
+        //find the mesh which has the skeleton attached to it
+        //this mesh will have a some size, which is needed for IntersectionMeshTrigger to happen
+        let otherMesh=AnimUtils.getMeshSkel(Vishva.vishva.avManager.avatar, true).mesh;
+        if (otherMesh == null) {
+            console.log("Cannot use this sensor as unable to find a mesh which as non zero size. The AV maynot have a skeleton");
+            return;
+        }
 
         if (properties.onEnter) {
             let action: Action = new ExecuteCodeAction({ trigger: ActionManager.OnIntersectionEnterTrigger, parameter: { mesh: otherMesh, usePreciseIntersection: false } }, (e) => { return this.emitSignal(e) });
@@ -58,13 +66,14 @@ export class SensorContact extends SensorAbstract {
             this.actions.push(action);
 
         }
-
         if (properties.onExit) {
             let action: Action = new ExecuteCodeAction({ trigger: ActionManager.OnIntersectionExitTrigger, parameter: { mesh: otherMesh, usePreciseIntersection: false } }, (e) => { return this.emitSignal(e) });
             this.mesh.actionManager.registerAction(action);
             this.actions.push(action);
         }
     }
+
+
 
     private findAV(scene: Scene): AbstractMesh {